refactor(network): 重构API接口和服务层实现
- 将ApiAiModelResponse类的字段访问权限从private改为public
- 重新整理ApiService接口中的import语句并优化接口方法排序
- 更新updateUserInfo和logout方法的返回类型为BaseResponse<Void>
- 调整searchUser方法的返回类型为BaseResponse<User>
- 修正getUnreadMessages接口路径为/api/chat/messages/unread
- 添加getFriendList接口方法获取好友列表
- 更新AppDatabase实体类配置，将FriendRequest替换为FriendRequestEntity
- 重构BaseResponse类实现，添加静态成功/错误方法和完整的getter/setter
- 完善ChangePwdRequest类的构造方法和getter/setter实现
- 修正ChangePwdResponse类的字段定义和访问方法
- 优化ChatActivity中的线程池管理和数据库操作逻辑
- 修复ChatActivity中好友头像加载和消息重发功能
- 重构ChatAdapter的消息绑定逻辑，优化头像加载和状态处理

+++ app/src/main/java/com/example/myfirstapplication/database/ChatDao.java
@@ -11,19 +11,19 @@ import java.util.List;
 
 @Dao
 public interface ChatDao {
-    // 插入消息并返回生成的主键 ID (long)
+    // 插入消息（返回自增ID）
     @Insert
     long insert(ChatMessage message);
 
-    // 更新指定 ID 消息的内容和状态
+    // 修复：参数类型改为Integer（匹配实体类的id类型）
     @Query("UPDATE messages SET content = :newContent, status = :newStatus WHERE id = :msgId")
-    void updateMessage(long msgId, String newContent, int newStatus);
+    void updateMessage(Integer msgId, String newContent, Integer newStatus);
 
-    // 查询某个好友的聊天记录 (用于列表显示)
+    // 查询好友聊天记录（LiveData）
     @Query("SELECT * FROM messages WHERE friendId = :fId ORDER BY timestamp ASC")
     LiveData<List<ChatMessage>> getMessagesByFriend(String fId);
 
-    // 根据friendId查询聊天记录（供UI展示）
+    // 查询好友聊天记录（同步）
     @Query("SELECT * FROM messages WHERE friendId = :friendId ORDER BY timestamp ASC")
     List<ChatMessage> getChatMessages(String friendId);
 }
